// src/features/ShrineEditor/components/EditorArea/components/tabs/history/components/HistoryEditForm/helpers/HistoryForm.types.ts
import type { CitationFormValues } from "../../../../../../../../../shared/citations/helpers/CitationSection.types";
import type { ImageFormValues } from "../../../../../../../../../shared/images/helpers/ImageSection.types";

export type HistoryFormValues = {
  eventDate: string;
  sortOrder: string;
  title: string;
  information: string;
  image: ImageFormValues;
  citations: CitationFormValues[];
};